requirements.txt
groq>=0.4.1
tavily-python==0.2.6
matplotlib==3.8.2
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
langchain>=0.1.0
langchain-groq==0.2.4
langchain-core>=0.3.33,<0.4.0 